+++ examples/stubbed_model/Makefile
@@ -1,0 +1,2 @@
+dist/main: main.py
+	pyinstaller --onefile --hidden-import="googleapiclient" $^
